+++ R/defunct.R
@@ -102,3 +102,31 @@ NULL
 unset_auth <- function() {
   .Defunct(new = "unset_key", package = "hereR")
 }
+
+#' HERE Geocoder API: Autocomplete
+#'
+#' Completes addresses using the HERE 'Geocoder Autocomplete' API.
+#'
+#' @references
+#' \href{https://developer.here.com/documentation/geocoder-autocomplete/dev_guide/topics/resource-suggest.html}{HERE Geocoder API: Autocomplete}
+#'
+#' @param addresses character, addresses to autocomplete.
+#' @param results numeric, maximum number of suggestions (Valid range: 1 and 20).
+#' @param url_only boolean, only return the generated URLs (\code{default = FALSE})?
+#'
+#' @return
+#' A \code{data.frame} object, containing the autocomplete suggestions for the addresses.
+#' @name autocomplete-defunct
+#' @usage autocomplete(addresses, results, url_only)
+#' @seealso \code{\link{hereR-defunct}}
+#' @keywords internal
+NULL
+
+#' @rdname hereR-defunct
+#' @section \code{autocomplete}:
+#' For \code{autocomplete}, use \code{\link{autosuggest}}.
+#'
+#' @export
+autocomplete <- function(addresses, results = 5, url_only = FALSE) {
+  .Defunct(new = "autosuggest", package = "hereR")
+}